Moved file storage-commons to src/. References #58.
