+++ frontend/src/Session/Login.jsx
@@ -17,7 +17,7 @@ import { authLogin } from '../API/login';
 import { useNavigate } from 'react-router-dom';
 import { useContext } from 'react';
 import AppContext from '../Contexts/AppContext';
-import { SetInLocalStorage } from '../../storage-commons'
+import { SetInLocalStorage } from '../storage-commons'
 
 const defaultTheme = createTheme();
