+++ frontend/src/storage-commons.js
